02-cf-quotes-postgres/pyproject.toml
[project]
name = "cf-quotes-postgres-02"
version = "1.0.0"
description = "Random Quotes with PostgreSQL - Cloud Foundry demo with Python"
requires-python = ">=3.11"
dependencies = [
    # FastAPI and server
    "fastapi",
    "uvicorn",
    
    # PostgreSQL database
    "psycopg2-binary",
    
    # Cloud Foundry service binding
    "cfenv",
]
